Refactor serial functions into module serialHat
